agenthub/pyproject.toml
[project]
name = "ni.agenthub"
version = "0.1.0"
description = "AgentHub control plane for AgentKit"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "ni.agentkit>=0.4.1",
]
authors = [
    { name = "Krix Tam", email = "krix.tam@qq.com" }
]

[project.scripts]
agenthub = "agenthub.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agenthub*"]
